day-06 | added transaction mode understanding | suprasanna

+++ src/main/java/com/customer/profile/dto/CustomerPreferenceResponse.java
@@ -3,7 +3,7 @@ package com.customer.profile.dto;
 import com.customer.profile.enums.CustomerSize;
 
 public record CustomerPreferenceResponse(
-    int preferenceId,
+    Long preferenceId,
     CustomerSize size
 ) {
